requirements.txt
absl-py==0.5.0
alabaster==0.7.8
antlr4-python3-runtime==4.8
asn1crypto==0.24.0
atari-py==0.2.6
atomicwrites==1.2.1
attrs==18.2.0
autobahn==17.10.1
Automat==0.6.0
Babel==2.4.0
bleach==2.1.2
blinker==1.4
box2d-py==2.3.8
cachetools==4.2.4
carla==0.9.13
cbor==1.0.0
certifi==2018.8.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cloud-init==22.2
cloudpickle==1.3.0
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cycler==0.10.0
Cython==0.29.32
dataclasses==0.8
decorator==4.1.2
distro-info===0.18ubuntu0.18.04.1
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.14
entrypoints==0.2.3.post1
enum34==1.1.6
future==0.16.0
gitdb==4.0.9
GitPython==3.1.18
glfw==2.5.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
gym==0.17.1
gym-minigrid==1.0.2
html5lib==0.999999999
httplib2==0.9.2
hydra-core==1.0.6
hyperlink==17.3.1
idna==2.7
imageio==2.4.1
imagesize==0.7.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==16.10.1
ipykernel==4.8.2
ipython==5.5.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.10
jsonpatch==1.16
jsonpickle==0.9.6
jsonpointer==1.10
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keras==2.10.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
language-selector==0.1
lockfile==0.12.2
lz4==0.10.1
Markdown==3.3.7
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.3
mock==2.0.0
more-itertools==4.3.0
mpi4py==2.0.0
mpyq==0.2.5
mujoco-maze==0.2.0
mujoco-py==2.0.2.0
munch==2.3.2
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
numpy==1.19.5
numpydoc==0.7.0
oauthlib==3.2.2
olefile==0.45.1
omegaconf==2.0.6
opencv-python==4.5.1.48
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.4.2
pathlib2==2.3.2
pathtools==0.1.2
pbr==4.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.3.0
pluggy==0.7.1
portpicker==1.2.0
probscale==0.2.3
promise==2.3
prompt-toolkit==1.0.15
protobuf==3.19.6
psutil==5.9.4
py==1.6.0
py-ubjson==0.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pycrypto==2.6.1
pygame==1.9.4
pyglet==1.5.0
Pygments==2.2.0
PyGObject==3.26.1
PyJWT==1.5.3
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.2.2
PySC2==2.0.2
pyserial==3.4
pytest==3.8.2
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.7.3
python-debian==0.1.32
python-snappy==0.5
PyTrie==0.2
pytz==2018.3
pyxdg==0.25
PyYAML==6.0
pyzmq==16.0.2
qrcode==5.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
roman==2.0.0
rsa==4.9
s2clientprotocol==4.6.1.68195.0
sacred==0.7.2
scipy==1.1.0
SecretStorage==2.3.1
sentry-sdk==1.11.1
service-identity==16.0.0
setproctitle==1.2.3
shortuuid==1.0.11
simplegeneric==0.8.1
six==1.11.0
sk-video==1.1.10
smmap==5.0.0
snakeviz==1.0.0
sos==4.3
Sphinx==1.6.7
ssh-import-id==5.7
stable-baselines3==1.0
systemd-python==234
tensorboard==2.4.0
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
terminado==0.7
testpath==0.3.1
torch==1.7.1
torchvision==0.8.2
tornado==5.1.1
tqdm==4.42.1
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.1
typing_extensions==4.1.1
u-msgpack-python==2.1
ubuntu-advantage-tools==27.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
wandb==0.13.6
wcwidth==0.1.7
webencodings==0.5
websocket-client==0.53.0
Werkzeug==2.0.3
whichcraft==0.5.2
wrapt==1.10.11
wsaccel==0.6.2
zipp==3.6.0
zope.interface==4.3.2
